chore: #86 코드리뷰 반영
