+++ Cherrish-iOS/Cherrish-iOS/Domain/Model/ProcedurePhase.swift
@@ -1,0 +1,14 @@
+//
+//  ProcedurePhase.swift
+//  Cherrish-iOS
+//
+//  Created by 송성용 on 1/15/26.
+//
+
+import Foundation
+
+enum ProcedurePhase {
+    case sensitive
+    case caution
+    case recovery
+}
