+++ Cherrish-iOS/Cherrish-iOS/Presentation/Global/Extension/ProcedurePhase+.swift
@@ -1,5 +1,5 @@
 //
-//  ProcedurePhase.swift
+//  ProcedurePhase+.swift
 //  Cherrish-iOS
 //
 //  Created by 송성용 on 1/15/26.
@@ -7,11 +7,7 @@
 
 import Foundation
 
-enum ProcedurePhase {
-    case sensitive
-    case caution
-    case recovery
-
+extension ProcedurePhase {
     var displayText: String {
         switch self {
         case .sensitive:
